==> scripts/13_Adequacy_RQ1.Rmd <==
---
title: "Demographic disparity in Aarhus shelter provisioning?"
author: "Adela Sobotkova"
date: "2025-01-14"
output: html_document
---
This script explores the spatial adequacy of civil defense amenities in aarhus, by comparing the capacity of shelters (supply of protection) with the overall population in Aarhus (demand on protection). It uses the streamlined SR, BDG and KOB shelter data to model supply and aggregated population and houses as proxies for population for demand. 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(googlesheets4)
library(sf)
library(mapview)
library(terra)
```
## Load public shelter (BDG and KOB) Data

### Wide BDG data
one unique shelter per row reflecting 1987 or 2024 status (depending on which coordinates you choose) from
```{r}
# Load Andreas complete data from Google and create a 1987 version (Wide sf dataset)
BDG <- readRDS("../output_data/BDG_andreas.rds")
glimpse(BDG)
```
### Long BDG data
shelters over time from https://docs.google.com/spreadsheets/d/1C4GEgq4UmEwx_Xi84FiNBrRowmmJHi3S191byjF-hWc/edit?gid=0#gid=0
```{r}
# All shelters sheet with long format where temporal changes are represented
BDG <- read_sheet("https://docs.google.com/spreadsheets/d/1C4GEgq4UmEwx_Xi84FiNBrRowmmJHi3S191byjF-hWc/edit?gid=0#gid=0",
                  range = "Ark1", 
                  col_types = "ddcdddcdccdc")

BDG  <- BDG %>% 
  filter(!is.na(Long) | !is.na(Lat) ) %>% 
  st_as_sf(coords = c("Long", "Lat"), crs = 4326) 

glimpse(BDG)

# Plot BDG in space by year of construction
m <- BDG %>% 
  group_by(BDnr) %>% 
  summarize(Startyear = min(Location_startdate), 
            Capacity = min(Capacity)) %>% 
  mapview(cex = "Capacity", zcol = "Startyear")
m
```

### Beginning snapshot - 1944 BDG data
Long data streamlined into unique records only when the CD network was established
```{r BDG-1944}
BDG %>% 
  filter(Location_startdate==1944) %>% 
  mapview(cex = "Capacity")

BDG %>% 
  filter(Location_startdate == 1944) %>% 
 # filter(Location_startdate > 1950 & Location_startdate < 1960) %>% 
  group_by(Type) %>% 
  tally()
```
When looking at shelter types, in 1944 all shelters in Aarhus are type III (297) or IIIA(4). Variation is only added in 1960s, with 16 type IV and 16 type IVA, and 2 type VI shelters (A and B respectively).
Some 95 shelters have unconfirmed Type in 1944.


### KOB data
We have 36 combined public shelters in the Gsheet: https://docs.google.com/spreadsheets/d/1kBkz2PdR_jMYIx5Y014oWMiVepXb4sQVx9fc30Vwn4A/edit?gid=0#gid=0 

```{r KOB}
KOB <- readRDS("../output_data/KOB_sf.rds")
mapview(KOB, cex = "Capacity", zcol = "Year")
```

### Sikringsum - Private shelters
```{r sikringsrum}
SR <- read_sf("../output_data/SR_sikringsrum.geojson")
SR <- SR %>% 
  st_transform(4326)
```


## Public and private shelter capacity
What percentage of population can public and private shelters in Aarhus actually accommodate? How does it compare to population of Aarhus at diferent times (1950-60, 1985,etc.) How many can be on the move at any time in 1960 and 1985?

How many people are in the homes in Aarhus?

According to the Aarhus population data (https://docs.google.com/spreadsheets/d/1cXwKflMb2A2NydSd07XomroW9yz7nG3evaT6Yn2_XDQ/edit?gid=0#gid=0), there are 51814 people in Aarhus in 1901 and 150879 in 1950. This is topped by 195152 in 1986.

### Public and private shelters capacity - in 1987
using wide data (based on 1987 and 2024 assessments)
```{r capacity-1987-all}
# Final capacity of all shelters altogether 
BDG <- readRDS("../output_data/BDG_andreas.rds")
paste0("There are ", sum(is.na(BDG$Final_Pub_Size)), " BDGs without indicated capacity")
paste0(" Otherwise, all BDGs built could altogether accommodate ",sum(BDG$Final_Pub_Size, na.rm = T), " people, while KOBs could absorb ", sum(KOB$Capacity, na.rm = T), " residents and finally SR (sikringsrum) could intake ", sum(SR$places, na.rm = T), " people")
 
paste0("The city and residents of Aarhus created between 1940 - 1989 some ", sum(is.na(BDG$Final_Pub_Size)) +sum(KOB$Capacity, na.rm = T)+sum(SR$places, na.rm = T)," places in public and private shelters")
```
There were altogether `sum(is.na(BDG$Final_Pub_Size)) +sum(KOB$Capacity, na.rm = T)+sum(SR$places, na.rm = T)` places in public and private shelters throughout Aarhus city from 1940 and 1989.

### Public shelters capacity in 1960 and 1989
using long data
```{r capacity-1960-all}
# Final capacity of all shelters altogether 
BDG <- read_sheet("https://docs.google.com/spreadsheets/d/1C4GEgq4UmEwx_Xi84FiNBrRowmmJHi3S191byjF-hWc/edit?gid=0#gid=0",
                  range = "Ark1", 
                  col_types = "ddcdddcdccdc")

library(janitor)

# 301 public (BDG) shelters finished in 1944
BDG %>% 
  filter(Location_startdate < 1961) %>% 
  group_by(BDnr) %>%  # to catch moving shelters
  summarize(last_move = max(Location_startdate),  # grab the latest move
            capacity = last(Capacity)) %>%   # grab the latest capacity
  summarize(total_cap = sum(capacity, na.rm = T))  # 15250 places in 301 locations

# 338 public (BDG) shelters in 1989, but capacity drops somewhat (movement, decommissioning)
BDG %>% 
  group_by(BDnr) %>%  # to catch moving shelters
  summarize(last_move = max(Location_startdate),  # grab the latest move
            capacity = last(Capacity)) %>%   # grab the latest capacity
  summarize(total_cap = sum(capacity, na.rm = T))  # 14150 places in 338 locations
# I need to discard the discontinued-by-1989 shelters 
```

### Private shelters capacity in 1960 and 1989
```{r}
SR
sum(SR$places, na.rm = T)
SR %>% 
  filter(year < 1961) %>% 
  summarize(total_cap = sum(places, na.rm = T))  # 52,451 places

SR %>% 
  filter(year < 1990) %>% 
  summarize(total_cap = sum(places, na.rm = T))  # 234,349 places by 1989;

SR %>% 
  filter(year < 1986) %>% 
  summarize(total_cap = sum(places, na.rm = T))  # 216,327 places in 1985
SR %>% distinct(places) %>% arrange(desc(places))
```

## Check adequacy : how does capacity of the shelters map against city population?

In 1960, some 15,250 people can be on the move and accommodated in public and 52,451 people in private shelters in Aarhus. 
In 1985, there are ~253,650 people in Aarhus.

In 1985, the number of shelter places in public places drops to 14,150 despite the increasing population. In both times, combined shelters accommodate additional 13,000 people.
Private shelters accommodated 216,327 people by 1985.
Private shelters can accommodate 2 - 2000 people

The method: 

* First calculate geography-wide percentage: ie. fraction of population by city district
* Second, calculate data-implied percentage: ie. number of people who fit in shelters within a given city district.
* Third, calculate demographic disparity score by gauging the difference between the data-implied percentage and the geography-wide percentage.

For full explanation of steps, see:
https://ui-research.github.io/sedt_documentation/sedt_algorithm.html
Other options are using Taleai's network approach.

### Use urban districts and population to assess equity/adequacy of shelters
City limits: I think there are many different conceptualisations. In terms of CD, they employed a term called a civil defence area, in Danish, Civilforsvarsområde, CFO. In our case this included Aarhus Købstad and Viby, Aaby, Vejlby-Risskov, Holme-Tranbjerg, Hasle-Skejby-Lisbjerg og Brabrand-Sdr. Aarslev kommuner. 

This was the local level, of course CD was also organised regionally and nationally. 
```{r disticts}
aarhus_districts
mapview(aarhus_districts, zcol = "prog_distrikt_navn")
```

### 1a. Estimate population by house floors

```{r estimate-population}
# Intersect aarhus districts with flats and estimate the houses' population
intersections_aarhus <- st_intersection(aarhus_districts, st_transform(bbr_aarhus_data_flat, 25832)) %>% 
  dplyr::select(id = id_lokalId, navn = prog_distrikt_navn, 
                byg021BygningensAnvendelse, byg026Year, byg054AntalEtager, byg406Koordinatsystem, city, geometry) %>% 
    mutate(byg021BygningensAnvendelse = as.numeric(byg021BygningensAnvendelse)) %>% 
  left_join(housecodes, by = c("byg021BygningensAnvendelse"="Code")) %>% 
  rename(Code = byg021BygningensAnvendelse) %>% 
  mutate(
    estimated_population = case_when(
      Purpose == "residential" & Code == "120" ~ 4,  # Single-family homes
      Purpose == "residential" & Code == "110" ~ 4,  # Single-family homes
      Purpose == "residential" & Code == "121" ~ 4,  # Single-family 
      Purpose == "residential" & Code == "122" ~ 4,  # Single-family homes
      Purpose == "residential" & Code %in% c("130", "131", "132") ~ 6,  # Single-family homes
      Purpose == "residential" & Code == "140" ~ 2 * byg054AntalEtager * 6,  # Multi-family, 2 units/floor
      Purpose == "residential" & Code == "150" ~ 2 * byg054AntalEtager * 6,  # Multi-family, 2 units/floor
      Purpose == "residential" & Code == "160" ~ 2 * byg054AntalEtager * 6,  # Multi-family, 2 units/floor
      Purpose == "residential" & Code == "190" ~ 2 * byg054AntalEtager ,  # unknown
      Purpose == "industry or production" & Code %in% c("220":"223") ~ 2 * byg054AntalEtager ,
      Purpose == "residential" & Code  %in% c("310":"314") ~ 2 * byg054AntalEtager *10,
      Purpose == "office" & Code  %in% c("320":"324") ~ 2 * byg054AntalEtager * 4,
      Purpose == "services" & Code  %in% c("330":"416") ~ 2 * byg054AntalEtager *4,
      Purpose == "educational" & Code  %in% c("419":"429") ~ 2 * byg054AntalEtager * 30,
      Purpose == "health services" & Code  %in% c("430":"439") ~ 2 * byg054AntalEtager * 15 ,
      
      TRUE ~ 0  # Non-residential or unknown
    )
  )


get_year_df <- function(df,year){
temp_df <- df %>% 
  filter(byg026Year %in% year)
    return(temp_df)
}
```

### 1b. Summarize population by district per year 
..or per an interval of years and new buildings.

```{r new-buildings-in-years}
## new buildings built in a certain interval
n_new_buildings_startend <- function(start_year, year) {
  n_new <- rep(0, length(aarhus_districts$prog_distrikt_navn))
  
  # Loop through each year from start_year to the specified year
  for (current_year in start_year:year) {
    intersections_year <- get_year_df(intersections_aarhus, current_year)
    
    # Aggregate data for each district
    for (i in 1:length(aarhus_districts$prog_distrikt_navn)){
      n_new[i] <- n_new[i] + nrow(filter(intersections_year, navn == aarhus_districts$prog_distrikt_navn[i]))
    }
  }

  temp <- cbind(aarhus_districts, n_new)
  return(temp)
}

buildings1900_1960 <- n_new_buildings_startend(1900, 1960)

## dplyr version for all buildings AND population up to a certain year

# Assuming intersections_aarhus already contains all data from all years
n_buildings_and_population_cumulative <- function(end_year) {
  # Filter data for all years up to the end year
  data_filtered <- intersections_aarhus %>%
    st_drop_geometry() %>% 
    filter(byg026Year <=end_year) %>%
    group_by(navn) %>%
    summarise(n_new = n(), # Count of new buildings
      total_population = sum(estimated_population, na.rm = TRUE)  # Sum of estimated population
    ) %>%
            
    ungroup()

  # Join with district names to ensure all districts are included
  result <- aarhus_districts %>%
    select(prog_distrikt_type_kode:prog_distrikt_nr) %>% 
    left_join(data_filtered, by = c("prog_distrikt_navn" = "navn"))

  # Replace NA with 0 in case some districts have no constructions
  result$n_new[is.na(result$n_new)] <- 0

  return(result)
}

buildings1960 <- n_buildings_and_population_cumulative(1960)
buildings1950 <- n_buildings_and_population_cumulative(1950)
mapview(buildings1960 %>% filter(n_new > 0), zcol = "total_population") + mapview(BDG, cex = 0.1)

buildings1960 %>% tally(total_population)
```
Second, calculate data-implied percentage: ie. number of people who fit in shelters within a given city district.
Third, calculate demographic disparity score by gauging the difference between the data-implied percentage and the geography-wide percentage.
For full explanation, see:
https://ui-research.github.io/sedt_documentation/sedt_algorithm.html

### 2. Calculate Implied Percentage : shelter population
```{r}
shelter_intersections <- aarhus_districts %>% 
   select(prog_distrikt_type_kode:prog_distrikt_nr) %>% 
  st_intersection(st_transform(BDG, 25832))

get_year_shelter <- function(df,year){
temp_df <- df %>% 
  filter(Location_startdate < year)
    return(temp_df)
}

shelter_capacity_district <- function(end_year) {
  # Filter data for all years up to the end year
  data_filtered <- shelter_intersections %>%
    st_drop_geometry() %>% 
    filter(Location_startdate <= 1950) %>%
    group_by(prog_distrikt_navn) %>%
    summarise(n_shelters = n(), # Count of shelters per district
      total_protected = sum(Capacity, na.rm = TRUE)  # Sum of estimated protected population
    ) %>%
            
    ungroup()

  # Join with district names to ensure all districts are included
  result <- aarhus_districts %>%
    select(prog_distrikt_type_kode:prog_distrikt_nr) %>% 
    left_join(data_filtered, by = "prog_distrikt_navn")

  # Replace NA with 0 in case some districts have no constructions
  result$n_new[is.na(result$n_shelters)] <- 0

  return(result)
}

shelter_capacity_district(1950) %>% 
  mapview(zcol = "n_shelters")

shelter_capacity_district(1950) %>% 
  mapview(zcol = "total_protected")
```

### Add KOB to total protected count
### Demographic disparity (adequacy)
Third, calculate demographic disparity score by gauging the difference between the data-implied percentage and the geography-wide percentage.
For full explanation, see:
https://ui-research.github.io/sedt_documentation/sedt_algorithm.html

```{r}
library(dplyr)

# Assume n_new_buildings_and_population_cumulative has been defined as above
# Assume shelter_capacity_district has been defined as provided

# Function to calculate demographic disparity scores
calculate_demographic_disparity <- function(end_year) {
  # Gather data for buildings and populations up to the specified year
  building_population_data <- n_buildings_and_population_cumulative(end_year)

  # Gather shelter capacity data up to the specified year
  shelter_data <- shelter_capacity_district(end_year) %>% st_drop_geometry()

  # Combine the datasets
  combined_data <- left_join(building_population_data, shelter_data, by = "prog_distrikt_navn")

  # Calculate disparity scores (example calculation)
  combined_data <- combined_data %>%
    mutate(
      demographic_disparity = round((total_population-total_protected) / total_population, 2),
      protected_pct = 1 - demographic_disparity
      # Example score calculation
    )

  # Replace NA with appropriate values in demographic_disparity
  combined_data$demographic_disparity[is.na(combined_data$demographic_disparity)] <- 0

  return(combined_data)
}

# Call the function for a specific year
disparity_scores <- calculate_demographic_disparity(1960)
mapview(disparity_scores, zcol = "protected_pct")

```
