==> scripts/14_Access_RQ1.Rmd <==
---
title: "Where are the gaps in Aarhus shelter provisioning?"
author: "Adela Sobotkova"
date: "2024-12-16"
output: html_document
---
This script explores the accessibility of the streamlined BDG and KOB shelters. Euclidean distances as well as network distances are calculated
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(googlesheets4)
library(sf)
library(mapview)
library(terra)
```
## Load Data
### BDG

#### Wide data
```{r}
# Load Andreas complete data from Google and create a 1987 version (Wide sf dataset)
BDG <- readRDS("../output_data/BDG_andreas.rds")
glimpse(BDG)
```
#### Long data
```{r}
# All shelters sheet with long format where temporal changes are represented
BDG <- read_sheet("https://docs.google.com/spreadsheets/d/1C4GEgq4UmEwx_Xi84FiNBrRowmmJHi3S191byjF-hWc/edit?gid=0#gid=0",
                  range = "Ark1", 
                  col_types = "ddcdddcdccdc")

BDG  <- BDG %>% 
  filter(!is.na(Long) | !is.na(Lat) ) %>% 
  st_as_sf(coords = c("Long", "Lat"), crs = 4326) 

glimpse(BDG)

# Plot BDG in space by year of construction
m <- BDG %>% 
  group_by(BDnr) %>% 
  summarize(Startyear = min(Location_startdate), 
            Capacity = min(Capacity)) %>% 
  mapview(cex = "Capacity", zcol = "Startyear")
m
```

#### 1944 data

```{r}

BDG %>% 
  filter(Location_startdate==1944) %>% 
  mapview(cex = "Capacity")

BDG %>% 
  filter(Location_startdate == 1944) %>% 
 # filter(Location_startdate > 1950 & Location_startdate < 1960) %>% 
  group_by(Type) %>% 
  tally()
```
When looking at shelter types, in 1944 all shelters in Aarhus are type III (297) or IIIA(4). Variation is only added in 1960s, with 16 type IV and 16 type IVA, and 2 type VI shelters (A and B respectively).
Some 95 lack Type.


### KOB data
We have 36 combined public shelters in the Gsheet: https://docs.google.com/spreadsheets/d/1kBkz2PdR_jMYIx5Y014oWMiVepXb4sQVx9fc30Vwn4A/edit?gid=0#gid=0 

```{r KOB}
KOB <- readRDS("../output_data/KOB_sf.rds")
mapview(KOB, cex = "Capacity", zcol = "Year")
```

### Sikringsum - Private shelters
```{r sikringsrum}
SR <- read_sf("../output_data/SR_sikringsrum.geojson")
SR <- SR %>% 
  st_transform(4326)
```


## Public and private shelter capacity
What percentage of population can public and private shelters in Aarhus actually accommodate? How does it compare to population of Aarhus at diferent times (1950-60, 1985,etc.) How many can be on the move at any time in 1960 and 1985?

How many people are in the homes in Aarhus?

According to the Aarhus population data (https://docs.google.com/spreadsheets/d/1cXwKflMb2A2NydSd07XomroW9yz7nG3evaT6Yn2_XDQ/edit?gid=0#gid=0), there are 51814 people in Aarhus in 1901 and 150879 in 1950. This is topped by 195152 in 1986.

### Public and private shelters capacity - altogether collated in time
```{r capacity-1987-all}
# Final capacity of all shelters altogether 
BDG <- readRDS("../output_data/BDG_andreas.rds")
paste0("There are ", sum(is.na(BDG$Final_Pub_Size)), " BDGs without indicated capacity")
paste0(" Otherwise, all BDGs built could altogether accommodate ",sum(BDG$Final_Pub_Size, na.rm = T), " people, while KOBs could absorb ", sum(KOB$Capacity, na.rm = T), " residents and finally SR (sikringsrum) could intake ", sum(SR$places, na.rm = T), " people")
 
paste0("The city and residents of Aarhus created between 1940 - 1989 some ", sum(is.na(BDG$Final_Pub_Size)) +sum(KOB$Capacity, na.rm = T)+sum(SR$places, na.rm = T)," places in public and private shelters")
```
There were altogether `sum(is.na(BDG$Final_Pub_Size)) +sum(KOB$Capacity, na.rm = T)+sum(SR$places, na.rm = T)` places in public and private shelters throughout Aarhus city from 1940 and 1989.

### Public shelters capacity in 1960 and 1989
```{r capacity-1960-all}
# Final capacity of all shelters altogether 
BDG <- read_sheet("https://docs.google.com/spreadsheets/d/1C4GEgq4UmEwx_Xi84FiNBrRowmmJHi3S191byjF-hWc/edit?gid=0#gid=0",
                  range = "Ark1", 
                  col_types = "ddcdddcdccdc")

library(janitor)

# 301 public (BDG) shelters finished in 1944
BDG %>% 
  filter(Location_startdate < 1961) %>% 
  group_by(BDnr) %>%  # to catch moving shelters
  summarize(last_move = max(Location_startdate),  # grab the latest move
            capacity = last(Capacity)) %>%   # grab the latest capacity
  summarize(total_cap = sum(capacity, na.rm = T))  # 15250 places in 301 locations

# 338 public (BDG) shelters in 1989, but capacity drops somewhat (movement, decommissioning)
BDG %>% 
  group_by(BDnr) %>%  # to catch moving shelters
  summarize(last_move = max(Location_startdate),  # grab the latest move
            capacity = last(Capacity)) %>%   # grab the latest capacity
  summarize(total_cap = sum(capacity, na.rm = T))  # 14150 places in 338 locations
```

### Private shelters capacity in 1960 and 1989
```{r}
SR
sum(SR$places, na.rm = T)
SR %>% 
  filter(year < 1961) %>% 
  summarize(total_cap = sum(places, na.rm = T))  # 52,451 places

SR %>% 
  filter(year < 1990) %>% 
  summarize(total_cap = sum(places, na.rm = T))  # 234,349 places by 1989;

SR %>% 
  filter(year < 1986) %>% 
  summarize(total_cap = sum(places, na.rm = T))  # 216,327 places in 1985
SR %>% distinct(places) %>% arrange(desc(places))
```
```{r}
mapview(aarhus_districts, zcol = "prog_distrikt_navn")
aarhus_districts
```

### Estimate population


## Euclidian proximity (200m)
One of the national directives for shelters was that residents should always have a shelter within 200m of their location, day and night. Let's see how the Aarhus public shelters fare vis-a-vis this plan?

### BDG and KOB Buffers
```{r}
bdg_buff <- st_buffer(st_transform(BDG, 25832), dist = 200)
kob_buff <- st_buffer(st_transform(KOB, 25832), dist = 200)

bdg_buff %>% 
  st_union() %>% 
  st_union(st_union(kob_buff)) %>% 
  st_as_sf() %>% 
  st_write("../output_data/public_buff200.geojson")

sh_buff200 <- read_sf("../output_data/public_buff200.geojson")

mapview(sh_buff200)
```

### Gaps in Euclidean buffers
I need the boundary of Aarhus to subtract these buffers from in order to see the gaps. Then I can plot/summarize the <1960 buildings in these gaps and gauge how many residents might have been at risk.

```{r}
```


## MapboxAPI proximity (5 mins)
Calculate paths to shelters for all Aarhus buildings before 1990. Which ones are missing out?
let's build a matrix of stacked 5-minute isochrones for all shelter points and then look how many buildings are in the gaps.
```{r accessibility-KOB-BDG, eval=FALSE}
token <- "pk.eyJ1IjoiYWRpdmVhIiwiYSI6ImNrcWdhNGlybjB4OG0ydnNjcWZtOG9mc3UifQ.EbNmMF9aF8th5cb-h5f1lQ"

BDG_walking_isos <- BDG %>% 
  #filter(Location_startdate <1960) %>% 
  mb_isochrone(profile = "walking", 
               time = 1:5, 
               id = "BDnr")

KOB_walking_isos <- KOB %>% 
  #filter(Year < 1960) %>% 
  mb_isochrone(profile = "walking", 
               time = 1:5, 
               id = "Number")

head(KOB_walking_isos) # 5 x 36 values
BDG_walking_isos # there is 5 x 498 values

saveRDS(BDG_walking_isos, "../output_data/BDG_walking_isos1_5.rds")
saveRDS(KOB_walking_isos, "../output_data/KOB_walking_isos1_5.rds")
```

## Walking accessibility in different decades 
For this we need to link the isochrones to shelter data, extract year of construction, map to and merge (st_union) by decade and then st_difference() from city boundary.
```{r walking-isos-1960-1980}
# Load isos
BDG_walking_isos <- readRDS("../output_data/BDG_walking_isos5.rds")
KOB_walking_isos <- readRDS("../output_data/KOB_walking_isos5.rds")


bdg1960 <- BDG_walking_isos %>% 
  filter(time == 5) %>% 
  filter(id %in% BDG$BDnr) %>% 
  cbind(st_drop_geometry(BDG)) %>% 
  filter(Location_startdate < 1960) %>% 
  st_union() %>% st_as_sf()

bdg1980 <- BDG_walking_isos %>% 
  filter(time == 5) %>% 
  filter(id %in% BDG$BDnr) %>% 
  cbind(st_drop_geometry(BDG)) %>% 
  filter(Location_startdate < 1980) %>% 
  st_union() %>% st_as_sf()


bdg1980diff <- BDG_walking_isos %>% 
    filter(time == 5) %>% 
  filter(id %in% BDG$BDnr) %>% 
  cbind(st_drop_geometry(BDG)) %>% 
  filter(Location_startdate > 1960 & Location_startdate < 1980) %>% 
  st_union() %>% st_as_sf()

kob1960 <- KOB_walking_isos %>% 
  filter(time == 5) %>% 
  cbind(st_drop_geometry(KOB)) %>% 
  filter(Year < 1960) %>% 
  st_union() %>% st_as_sf()

kob1980 <- KOB_walking_isos %>% 
  filter(time == 5) %>% 
  cbind(st_drop_geometry(KOB)) %>% 
  filter(Year < 1980) %>% 
  st_union() %>% st_as_sf()

kob1980diff <- KOB_walking_isos %>% 
  filter(time == 5) %>% 
  cbind(st_drop_geometry(KOB)) %>% 
  filter(Year > 1960 & Year < 1980) %>% 
  st_union() %>% st_as_sf()

# Situation in 1960
mapview(bdg1960) + mapview(kob1960)

# Situation in 1980
mapview(bdg1980) + mapview(kob1980)

# growth in between
mapview(bdg1980diff) + mapview(kob1980diff)
```

### Matrix of accessibility
Now that we have the one-to-five-minute isochrones, let's overlay them and create an accessibility matrix.

But first we need to link our 1:5 minute isochrones to attribute data (which we need to duplicate 5 times for each 1 min isochrone)
```{r}
# Duplicate each row 5 times
pentupled_BDG <- BDG %>%
  st_drop_geometry() %>% 
  uncount(5, .id = NULL)  # .id = NULL prevents the creation of an id column

pentupled_KOB <- KOB %>%
  st_drop_geometry() %>% 
  uncount(5, .id = NULL)  # .id = NULL prevents the creation of an id column

# View the result
print(pentupled_BDG) # 2490 values
print(pentupled_KOB)  # 180 values
```
Now we can link the pentupled KOB and BDG attributes to their 1:5 min isochrones
and select locations that predate 1960s (preparedness in light of the bomb)
```{r}
BDG_isos1960_prj <- BDG_walking_isos %>% 
  cbind(pentupled_BDG) %>% 
  filter(Location_startdate < 1960) %>% 
  st_transform(25832)

KOB_isos1960_prj <- KOB_walking_isos %>% 
  cbind(pentupled_KOB) %>% 
  filter(Year < 1960) %>% 
  st_transform(25832)

```

### Create an accessibility surface out of the 1:5 isochrones to 1960s locations
First, we create separate surfaces with 50m resolution for each KOB and BDG category
```{r}
library(Rcpp)
library(fasterize)
library(raster)
library(leaflet)

# Create rasters from BDG and KOB 1:5min 1960 location isochrones
template <- raster(BDG_isos1960_prj, resolution = 50)
BDG1960_iso_surface <- fasterize(BDG_isos1960_prj, template, field = "time", fun = "min")

#writeRaster(BDG1960_iso_surface, "../output_data/BDG1960_iso_surf.tiff", format = 'GTiff' )

template_kob <- raster(KOB_isos1960_prj, resolution = 50)
KOB1960_iso_surface <- fasterize(KOB_isos1960_prj, template_kob, field = "time", fun = "min")

# Map the results
pal <- colorNumeric("viridis", BDG_isos1960_prj$time, na.color = "transparent")

mapbox_map <- leaflet() %>%
    addMapboxTiles(style_id = "streets-v11", username = "mapbox")

mapbox_map %>%
    addRasterImage(BDG1960_iso_surface, colors = pal, opacity = 0.5) %>%
    addLegend(values = BDG_isos1960_prj$time, pal = pal, title = "Walktime to the nearest BDG shelter")

mapbox_map %>%
    addRasterImage(KOB1960_iso_surface, colors = pal, opacity = 0.5) %>%
    addLegend(values = KOB_isos1960_prj$time, pal = pal, title = "Walktime to the nearest KOB shelter")

```
### Accessibility surface for 1:5 min walking distances to combined BDG and KOB
Next, we combine the BDG and KOB isochrones and let the fasterize function select the lowest value for each cell out of the many overlapping isochrones. This captures the 'best scenario', which is unlikely, however captures the planners' best intentions.
```{r public-shelter-surface}
BDG_KOB_iso1960 <- BDG_isos1960_prj %>% 
  select(time, id, Location_startdate) %>% 
  rename(Year = Location_startdate) %>% 
  rbind(KOB_isos1960_prj %>% select(time, id, Year))

template_public <- raster(BDG_KOB_iso1960, resolution = 50)
BDG_KOB_60s_iso <- fasterize(BDG_KOB_iso1960, template_public, field = "time", fun = "min")

#writeRaster(BDG_KOB_60s_iso, "../output_data/BDG_KOB_60iso_surf.tiff", format = 'GTiff' )

mapbox_map <- leaflet() %>%
    addMapboxTiles(style_id = "light-v10", username = "mapbox")

mapbox_map %>%
    addRasterImage(BDG_KOB_60s_iso, colors = pal, opacity = 0.5) %>%
    addLegend(values = BDG_KOB_iso1960$time, pal = pal, title = "Walktime to the nearest shelter")

```

### Plot accessibility : combined BDG and KOB isochrones
Let's first look at all the shelters' buffers in relation to the Aarhus 1960s flats within the city wider boundary.
```{r plot-flats-shelter-isos}
BDG_walking_isos <- readRDS("../output_data/BDG_walking_isos1_5.rds")
KOB_walking_isos <- readRDS("../output_data/KOB_walking_isos1_5.rds")
BDG_KOB_60s_iso <- rast("../output_data/BDG_KOB_60iso_surf.tiff" )
towns <- read_sf("../data/large/DanishCities2024.geojson")
aarhus <- towns %>% 
  filter(bebyggelseskode == 11045) 

# plot(BDG_KOB_60s_iso)
# mapview(BDG_KOB_60s_iso)

bbr_aarhus_data_flat %>% 
  filter(byg026Year < 1961) %>% 
  mapview(zcol = "byg054AntalEtager",
        cex = 0.1) + 
mapview(aarhus$geometry)+ 
mapview(BDG_walking_isos %>% filter(time == 5))
```

Now, lets clip the flats by city boundary and by shelter isos, so we see the ones without protection. Repeat with business and institutional buildings

```{r flats-1960-boundary}
# Select residences built by 1960 roughly within 2024 Aarhus boundary (excessive I know) 
aarhus_flats <- bbr_aarhus_data_flat %>%   # 50,000 flats
  filter(byg026Year < 1961) %>% # 16473 residencies built before 1960
  st_intersection(aarhus$geometry) # 14283 within Aarhus wider boundary

# Merge public shelters into single multipolygons to find overlapping flats
BDG_uni <- BDG_walking_isos %>% 
  filter(time == 5) %>% 
  st_union() %>% 
  st_make_valid()
KOB_uni <- KOB_walking_isos %>% 
  filter(time == 5) %>% 
  st_union() %>% 
  st_make_valid()

st_is_valid(KOB_uni)

# View the merged result == spatial coverage by public shelters 
mapview(KOB_uni) + mapview(BDG_uni)

```
### Flats outside the 5 min walking distance and their estimated population
```{r spatial-equity-residences-5min-1960}
# View flats (<1960) outside the 5min walking interval from public shelters
aarhus_flats %>% 
  st_difference(st_union(BDG_uni, KOB_uni)) %>% 
  mapview(zcol = "nearestBDG_dist_m", cex = "byg054AntalEtager")

# Residential houses outside 5 min walking isochrones from public shelters in 1960s Aarhus
outside5isos1960 <- aarhus_flats %>% 
  st_difference(st_union(BDG_uni, KOB_uni))  # only 3462 flats are unprotected / well outside of town!
```





## FLATS COVERAGE 1980s TO BE COMPLETED REDO for BDGS confirmed in 1985
Most shelter records we have are from 1987. Let's check what the spatial coverage was for 1985.
```{r public-sh-coverage-1984}
aarhus_flats1985 <- bbr_aarhus_data_flat %>%   # 50,000 flats
  filter(byg026Year <= 1985) %>% # 16473 residencies built before 1960
  st_intersection(aarhus$geometry)

# Residential houses outside 5 min walking isochrones from public shelters in 1960s Aarhus
outside5isos1985 <- aarhus_flats1985 %>% 
  st_difference(st_union(BDG_uni, KOB_uni))

outside5isos1985 %>% mapview(zcol = "nearestBDG_dist_m", cex = "byg054AntalEtager")
summary(outside5isos1985)
```
A whooping 13351  flats are outside 5 min walking interval to a shelter by 1985. Most of these however are outside the immediate danger zone (downtown), where only a dozen new buildings have popped up in the shelter gaps.

### Map proper of the gaps
```{r}
mapview(outside5isos1960, zcol = "nearestBDG_dist_m", cex = "byg054AntalEtager")
mapview(outside5isos1985, zcol = "nearestBDG_dist_m", cex = "byg054AntalEtager")
```


Most of the other 13000 homes are single-floor family houses, expected to be furnished with private shelters. Only one 17-floor building is in this list. 

## Estimate population for flats by type FINISH
```{r estimate-population}
# Intersect aarhus districts with flats and estimate the houses' population
intersections_aarhus <- st_intersection(aarhus_districts, st_transform(bbr_aarhus_data_flat, 25832)) %>% 
  dplyr::select(id = id_lokalId, navn = prog_distrikt_navn, 
                byg021BygningensAnvendelse, byg026Year, byg054AntalEtager, byg406Koordinatsystem, city, geometry) %>% 
    mutate(byg021BygningensAnvendelse = as.numeric(byg021BygningensAnvendelse)) %>% 
  left_join(housecodes, by = c("byg021BygningensAnvendelse"="Code")) %>% 
  rename(Code = byg021BygningensAnvendelse) %>% 
  mutate(
    estimated_population = case_when(
      Purpose == "residential" & Code == "120" ~ 4,  # Single-family homes
      Purpose == "residential" & Code == "110" ~ 4,  # Single-family homes
      Purpose == "residential" & Code == "121" ~ 4,  # Single-family 
      Purpose == "residential" & Code == "122" ~ 4,  # Single-family homes
      Purpose == "residential" & Code %in% c("130", "131", "132") ~ 6,  # Single-family homes
      Purpose == "residential" & Code == "140" ~ 2 * byg054AntalEtager * 6,  # Multi-family, 2 units/floor
      Purpose == "residential" & Code == "150" ~ 2 * byg054AntalEtager * 6,  # Multi-family, 2 units/floor
      Purpose == "residential" & Code == "160" ~ 2 * byg054AntalEtager * 6,  # Multi-family, 2 units/floor
      Purpose == "residential" & Code == "190" ~ 2 * byg054AntalEtager ,  # unknown
      Purpose == "industry or production" & Code %in% c("220":"223") ~ 2 * byg054AntalEtager ,
      Purpose == "residential" & Code  %in% c("310":"314") ~ 2 * byg054AntalEtager *10,
      Purpose == "office" & Code  %in% c("320":"324") ~ 2 * byg054AntalEtager * 4,
      Purpose == "services" & Code  %in% c("330":"416") ~ 2 * byg054AntalEtager *4,
      Purpose == "educational" & Code  %in% c("419":"429") ~ 2 * byg054AntalEtager * 30,
      Purpose == "health services" & Code  %in% c("430":"439") ~ 2 * byg054AntalEtager * 15 ,
      
      TRUE ~ 0  # Non-residential or unknown
    )
  )
```

## Outside SR isos FINISH
Check the protection of the 3000 and 13000 homes outside BDGisos with e.g. private shelters


```{r sirkingsum-outof5}
SR %>% 
  filter(year < 1961) %>% 
mapview(cex = "places")+ mapview(outside5isos1960, cex = 0.1)

SR %>% 
  #filter(year < 1961) %>% 
mapview(cex = "places")+ mapview(outside5isos1985, cex = 0.1)
```





### Compare Euclidian with Network distance to shelters
Let's compare the 200m buffer with 2 minute network distance
```{r}
# filter buffers to 1960s only
BDG_KOB_60s_buff200 <- bdg_buff %>% 
  filter(Location_startdate < 1960) %>% 
  st_union() %>% 
  st_union(st_union(kob_buff %>% filter(Year < 1960))) %>% 
  st_as_sf() 
  
st_write(BDG_KOB_60s_buff200,"../output_data/BDG_KOB_60s_buff200.geojson")


mapview(BDG_KOB_60s_buff200) + mapview(BDG_KOB_iso1960 %>% filter(time < 4), zcol = "time") 

mapview(BDG_KOB_60s_iso) + mapview(BDG_KOB_60s_buff200)
```
Finally, once we see what the gaps are, we can explore the buildings and population inside.
